Add function to calculate sigma for 68% CI
When we use a t-distribution model, we can't directly compare the
results for sigma to those derived from the normal model, because a
model with a smaller nu value will have a correspondingly smaller sigma
value. We can, however, calculate a scale factor for the 68% CI, which
allows us to directly compare with the sigma in the normal model.