Update the prior on intrinsic scatter
I changed the prior following some of the discussion in the Stan prior
choice recommendations, which referenced Chung et al. (2013,
Psychometrika). The reasoning behind this is that heavy weight on zero
intrinsic scatter can produce biased inference (in that the scatter is
biased to be too low). The recommendation in the paper is gamma(2, eps)
for eps -> 0, but this gives quite a broad prior... when we'd actually
expect sigma_int to be, at most, about 1, because we've scaled all of
the data to have zero mean and unit variance. Hence, I've gone for a
weakly informative prior of gamma(2, 2).